Move network mapper to protocol
