+++ mikeio1d/experimental/_network_mapper.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 
-import networkx as nx
 import pandas as pd
 
 from pathlib import Path
@@ -16,7 +15,6 @@ from mikeio1d.experimental._network_protocol import (
     NetworkEdge,
     NetworkNode,
     EdgeBreakPoint,
-    GenericNetwork,
     NetworkNodeIdGenerator,
 )
 
@@ -176,211 +174,3 @@ def parse_res1d_network(res: Any, id_gen: Callable[..., str]) -> dict[str, Netwo
 
     network = read_res1d_network(res)
     return {reach_id: parse_reach(reach) for reach_id, reach in network.reaches.items()}
-
-
-class NetworkMapper:
-    """Mapper class to transform Res1D to a general network coord system."""
-
-    def __init__(self, edges_dict: dict[str, NetworkEdge], id_generator: NetworkNodeIdGenerator):
-        self._alias_map: dict[int, str] = {}
-        # We include id_generator because we need a way to generate unique ids
-        # of elements with arbitrary architecture. For instance, here we might pass
-        # a function that depends on 'node' and 'reach' from res1d
-        self._id_generator = id_generator
-        self._edges = edges_dict
-
-    def map_network(self) -> GenericNetwork:
-        """Return generic network object.
-
-        Returns
-        -------
-        GenericNetwork
-        """
-        g0 = self._initialize_graph()
-        self._alias_map = {g0.nodes[id]["alias"]: id for id in g0.nodes()}
-
-        return GenericNetwork(g0)
-
-    def _initialize_graph(self) -> nx.Graph:
-        g0 = nx.Graph()
-        for edge in self._edges.values():
-            # 1) Add start and end nodes
-            for node in [edge.start, edge.end]:
-                if node.id in g0.nodes:
-                    g0.nodes[node.id]["boundary"].update(node.boundary)
-                else:
-                    g0.add_node(
-                        node.id,
-                        data=node.data,
-                        boundary=node.boundary,
-                    )
-
-            # 2) Add edges connecting start/end nodes to their adjacent breakpoints
-            if edge.n_breakpoints == 0:
-                g0.add_edge(edge.start.id, edge.end.id, length=edge.length)
-            else:
-                for breakpoint in edge.breakpoints:
-                    g0.add_node(breakpoint.id, data=breakpoint.data)
-
-                g0.add_edge(
-                    edge.start.id, edge.breakpoints[0].id, length=edge.breakpoints[0].distance
-                )
-
-                g0.add_edge(
-                    edge.breakpoints[-1].id,
-                    edge.end.id,
-                    length=edge.length - edge.breakpoints[-1].distance,
-                )
-
-            # 3) Connect consecutive intermediate breakpoints
-            for i in range(edge.n_breakpoints - 1):
-                current_ = edge.breakpoints[i]
-                next_ = edge.breakpoints[i + 1]
-                length = next_.distance - current_.distance
-                g0.add_edge(current_.id, next_.id, length=length)
-
-        return nx.convert_node_labels_to_integers(g0, label_attribute="alias")
-
-    def find(
-        self,
-        node: str | list[str] | None = None,
-        edge: str | list[str] | None = None,
-        distance: str | float | list[str | float] | None = None,
-    ) -> int | list[int]:
-        """Find node or breakpoint id in the generic network.
-
-        Parameters
-        ----------
-        node : str | List[str], optional
-            Node id(s) in the original network, by default None
-        edge : str | List[str], optional
-            Edge id(s) for breakpoint lookup or edge endpoint lookup, by default None
-        distance : str | float | List[str | float], optional
-            Distance(s) along edge for breakpoint lookup, or "start"/"end"
-            for edge endpoints, by default None
-
-        Returns
-        -------
-        int | List[int]
-            Node or breakpoint id(s) in the generic network
-
-        Raises
-        ------
-        ValueError
-            If invalid combination of parameters is provided
-        KeyError
-            If requested node/breakpoint is not found in the network
-        """
-        # Determine lookup mode
-        by_node = node is not None
-        by_breakpoint = edge is not None or distance is not None
-
-        if by_node and by_breakpoint:
-            raise ValueError(
-                "Cannot specify both 'node' and 'edge'/'distance' parameters simultaneously"
-            )
-
-        if not by_node and not by_breakpoint:
-            raise ValueError("Must specify either 'node' or both 'edge' and 'distance' parameters")
-
-        if by_node:
-            # Handle node lookup
-            if not isinstance(node, list):
-                node = [node]
-            ids = [self._id_generator(node_i) for node_i in node]
-
-        else:
-            # Handle breakpoint/edge endpoint lookup
-            if edge is None or distance is None:
-                raise ValueError(
-                    "Both 'edge' and 'distance' parameters are required for breakpoint/endpoint lookup"
-                )
-
-            if not isinstance(edge, list):
-                edge = [edge]
-
-            if not isinstance(distance, list):
-                distance = [distance]
-
-            # We can pass one edge and multiple breakpoints/endpoints
-            if len(edge) == 1:
-                edge = edge * len(distance)
-
-            if len(edge) != len(distance):
-                raise ValueError(
-                    "Incompatible lengths of 'edge' and 'distance' arguments. One 'edge' admits multiple distances, otherwise they must be the same length."
-                )
-
-            ids = []
-            for edge_i, distance_i in zip(edge, distance):
-                if distance_i in ["start", "end"]:
-                    # Handle edge endpoint lookup
-                    if edge_i not in self._edges:
-                        raise KeyError(f"Edge '{edge_i}' not found in the network.")
-
-                    network_edge = self._edges[edge_i]
-                    if distance_i == "start":
-                        ids.append(network_edge.start.id)
-                    else:  # distance_i == "end"
-                        ids.append(network_edge.end.id)
-                else:
-                    # Handle breakpoint lookup
-                    ids.append(self._id_generator(edge=edge_i, distance=distance_i))
-
-        # Check if all ids exist in the network
-        alias_set = set(self._alias_map.keys())
-        if all([id in alias_set for id in ids]):
-            if len(ids) == 1:
-                return self._alias_map[ids[0]]
-            else:
-                return [self._alias_map[id] for id in ids]
-        else:
-            missing_ids = [id for id in ids if id not in alias_set]
-            raise KeyError(
-                f"Node/breakpoint(s) {missing_ids} not found in the network. Available nodes are {alias_set}"
-            )
-
-    def recall(self, id: int | list[int]) -> dict[str, Any] | list[dict[str, Any]]:
-        """Recall the original coordinates from generic network node id(s).
-
-        Parameters
-        ----------
-        id : int | List[int]
-            Node id(s) in the generic network
-
-        Returns
-        -------
-        Dict[str, Any] | List[Dict[str, Any]]
-            Original coordinates. For single input returns dict, for multiple inputs returns list of dicts.
-            Dict contains coordinates:
-            - For nodes: 'node' key with node id
-            - For breakpoints: 'edge' and 'distance' keys with edge id and distance
-
-        Raises
-        ------
-        KeyError
-            If node id is not found in the network
-        ValueError
-            If node id string format is invalid
-        """
-        # Convert to list for uniform processing
-        if not isinstance(id, list):
-            id = [id]
-
-        # Create reverse lookup map
-        reverse_alias_map = {v: k for k, v in self._alias_map.items()}
-
-        results = []
-        for node_id in id:
-            if node_id not in reverse_alias_map:
-                raise KeyError(f"Node ID {node_id} not found in the network.")
-
-            string_id = reverse_alias_map[node_id]
-            coordinates = self._id_generator.parse(string_id)
-            results.append(coordinates)
-
-        # Return single dict if single input, list otherwise
-        if len(results) == 1:
-            return results[0]
-        else:
-            return results
